@@ -22,7 +22,8 @@ except ImportError:
     raise SystemExit(dedent(message)) from None
 
 package = "fagfunksjoner"
-python_versions = ["3.10", "3.11", "3.12"]
+python_versions = ["3.11", "3.12", "3.13"]
+python_versions_for_test = python_versions + ["3.10"]
 nox.needs_version = ">= 2021.6.6"
 nox.options.sessions = (
     "pre-commit",
@@ -152,7 +153,7 @@ def mypy(session: Session) -> None:
         session.run("mypy", f"--python-executable={sys.executable}", "noxfile.py")
 
 
-@session(python=python_versions)
+@session(python=python_versions_for_test)
 def tests(session: Session) -> None:
     """Run the test suite."""
     session.install(".")
